refactor: move space type
